Refactor action

+++ src/main/java/com/example/ics/services/impl/DataAccessServiceImpl.java
@@ -1,12 +1,10 @@
 package com.example.ics.services.impl;
 
 import com.example.ics.exceptions.ImageNotFoundException;
-import com.example.ics.exceptions.exception_handlers.ExceptionMessage;
 import com.example.ics.models.dtos.image.OriginType;
-import com.example.ics.models.dtos.image.ReadImageDto;
+import com.example.ics.models.dtos.image.ImageDto;
 import com.example.ics.models.dtos.tag.TagDto;
 import com.example.ics.models.dtos.image.PersistImageDto;
-import com.example.ics.models.dtos.image.UpdateImageDto;
 import com.example.ics.models.entities.Image;
 import com.example.ics.models.entities.Tag;
 import com.example.ics.repositories.ImageRepository;
@@ -23,7 +21,6 @@ import org.springframework.stereotype.Service;
 import java.util.*;
 import java.util.stream.Collectors;
 
-import static com.example.ics.exceptions.exception_handlers.ExceptionMessage.*;
 
 @Service
 public class DataAccessServiceImpl implements DataAccessService {
@@ -39,75 +36,99 @@ public class DataAccessServiceImpl implements DataAccessService {
 
     @Override
     @Transactional
-    public ReadImageDto getImageForReadByUrl(String imageUrl) {
+    public ImageDto getImageForReadByUrl(String imageUrl) {
         Image image = getImageByUrl(imageUrl);
         if (image == null) {
             return null;
         }
 
-        return new ReadImageDto(image, OriginType.FETCHED);
+        return new ImageDto(image, OriginType.FETCHED);
     }
 
     @Override
     @Transactional
-    public ReadImageDto persist(PersistImageDto imageDto, Set<TagDto> relatedTagDtos) {
+    public ImageDto getImageForReadById(String imageId) throws ImageNotFoundException {
+        Image image = getImageById(imageId);
+        if (image == null) {
+            throw new ImageNotFoundException();
+        }
+
+        return new ImageDto(image, OriginType.FETCHED);
+    }
+
+    @Override
+    public ImageDto getImageForReadByChecksum(String checksum) {
+        Optional<Image> optImage = imageRepository.findByChecksum(checksum);
+
+        if (optImage.isEmpty()) {
+            return null;
+        }
+
+        return new ImageDto(optImage.get(), OriginType.FETCHED);
+    }
+
+    @Override
+    @Transactional
+    public ImageDto persist(PersistImageDto imageDto, Set<TagDto> relatedTagDtos) {
         Set<Tag> tagEntities = transformTagDtosToEntities(relatedTagDtos);
         imageRepository.saveAndFlush(new Image(imageDto, tagEntities));
-        return new ReadImageDto(getImageByUrl(imageDto.url()), OriginType.CREATED);
+        return new ImageDto(getImageByUrl(imageDto.url()), OriginType.CREATED);
     }
 
 
     @Override
     @Transactional
-    public ReadImageDto update(ReadImageDto imageDto, Set<TagDto> relatedTagDtos) throws ImageNotFoundException {
-        Image imageForUpdate = imageRepository.findByUrl(imageDto.getUrl())
-                .orElseThrow(() -> new ImageNotFoundException(NOT_FOUND_IMAGE));
+    public ImageDto updateTagsById(String imageId, Set<TagDto> newTagDtos) {
+        Image imageForUpdate = getImageById(imageId);
+
+        if (imageForUpdate == null) {
+            return new ImageDto();
+        }
 
         Set<Tag> oldTags = imageForUpdate.getTags();
         tagRepository.deleteAll(oldTags);
 
-        Set<Tag> newTags = transformTagDtosToEntities(relatedTagDtos);
+        Set<Tag> newTags = transformTagDtosToEntities(newTagDtos);
 
-        imageForUpdate.setTags(newTags);
-        imageForUpdate.setAnalysedAt();
+        imageForUpdate.updateTags(newTags);
 
         imageRepository.saveAndFlush(imageForUpdate);
 
-        return new ReadImageDto(imageForUpdate, OriginType.CREATED);
+        return new ImageDto(imageForUpdate, OriginType.CREATED);
     }
+
 
     @Override
     @Transactional
-
-    public ReadImageDto getImageForReadById(String imageId) throws ImageNotFoundException {
-        Optional<Image> optImage = imageRepository.findById(imageId);
-        if (optImage.isEmpty()) {
-            throw new ImageNotFoundException(NOT_FOUND_IMAGE);
-        }
-
-        return new ReadImageDto(optImage.get(), OriginType.FETCHED);
-    }
-
-    @Override
-    @Transactional
-    public List<ReadImageDto> getPageOfImagesForReadBy
-            (boolean ascOrder, int pageNum, int pageSize, List<String> tagNames) {
+    public List<ImageDto> getPageOfImagesForReadBy
+            (boolean ascOrder, int pageNum, int pageSize, List<String> tagNames)
+            throws ImageNotFoundException {
 
         Pageable pageable = getImagePageOf(ascOrder, pageNum, pageSize);
         Page<Image> page;
+
         if (tagNames != null) {
             page = imageRepository.findAllThatContain(tagNames, pageable);
         } else {
             page = imageRepository.findAll(pageable);
         }
 
+        if (page.isEmpty()) {
+            throw new ImageNotFoundException();
+        }
+
         return getImagesForReadOutOf(page);
     }
 
+    @Override
+    public void deleteImage(String id) {
+        imageRepository.deleteById(id);
+    }
 
-    private List<ReadImageDto> getImagesForReadOutOf(Page<Image> page) {
+
+    private List<ImageDto> getImagesForReadOutOf(Page<Image> page) {
         return page.stream()
-                .map(entity -> new ReadImageDto(entity, OriginType.CREATED))
+                .map(entity -> new ImageDto(entity, OriginType.CREATED))
                 .toList();
     }
 
@@ -127,17 +148,14 @@ public class DataAccessServiceImpl implements DataAccessService {
         return imageRepository.findByUrl(url).orElse(null);
     }
 
+    private Image getImageById(String id) {
+        return imageRepository.findById(id).orElse(null);
+    }
+
     private Set<Tag> transformTagDtosToEntities(Set<TagDto> tagDtos) {
         return tagDtos
                 .stream()
                 .map(Tag::new)
-                .collect(Collectors.toCollection(TreeSet::new));
-    }
-
-    private Set<Tag> transformTagMapToEntities(Map<String, Integer> mapOfTags) {
-        return mapOfTags.entrySet()
-                .stream()
-                .map(e -> new Tag(e.getKey(), e.getValue()))
                 .collect(Collectors.toCollection(TreeSet::new));
     }
